add prometheus-operator/kube-prometheus manifests at v0.9.0 as submodule
Argo CD has [support for git submodules] and this allows us to pin the
manifests to a particular git sha (for v0.9.0) with confidence.  We could
refer to the manifests using git sha (or tag...if we trust it not to move)
in the upstream repo, but this feels a bit cleaner.  And we'll bump the
subproject commit for `vendor/kube-prometheus` when we're ready to
upgrade.

[support for git submodules]: https://github.com/argoproj/argo-cd/pull/2495