database: use preset random password for pg12
Previously the `bitnami/postgresql` helm template was
picking a random password.  That password ends up on the
persistent data volume and can fall out-of-sync in
circumstances where helm resets the k8s secret.  When
that happens, the backup jobs will fail because their
password from the k8s secret is no longer honored.

Avoid the out-of-sync condition by just generating a
password and manually creating the k8s secret.