database: enable TLS for pg12 connections
First step on the way to using mTLS for client connections.
`step certificate bundle` is used to create the chain
expected by Postgres for an intermediate-signed server
cert.  That's per
https://www.postgresql.org/docs/12/ssl-tcp.html#SSL-CERTIFICATE-CREATION :

> It is also possible to create a chain of trust that includes
> intermediate certificates:
> ...
> `server.crt` and `intermediate.crt` should be concatenated into
> a certificate file bundle and stored on the server.

Note that the server will log errors like this under normal
operation:

    2022-01-20 17:05:53.611 GMT [141] LOG:  provided user name (postgres) and authenticated user name (pg12) do not match
    2022-01-20 17:05:53.611 GMT [141] FATAL:  certificate authentication failed for user "postgres"
    2022-01-20 17:05:53.611 GMT [141] DETAIL:  Connection matched pg_hba.conf line 1: "hostssl     all             all             0.0.0.0/0               cert"

They happen because the default bitnami `pg_isready` health
probes attempt to connect to the server with user `postgres`
but using the server certificate with subject common name
`pg12`.

The errors can probably be eliminated by customizing the helm
values to mount the `postgres` cert and key and using custom
health probes.  For now, just planning to ignore them.