database: use `postgres` client cert for health checks
Re-introduce the `postgres` client certificate that was removed in
previous commit as a cert-manager `Certificate` resource.  Also
configure the liveness and readiness probes to use the client cert
for the `pg_isready` checks.

Unfortunately, https://github.com/kubernetes/kubernetes/issues/57923
turned this into a yak shave because the private key for the certificate
cannot be mounted with mode 0600.  Hence the new init container that
mounts the secret and copies it to a volume with the correct permissions.
This is the same pattern as used in bitnami/charts#3057 for the PG server
certs.