monitoring: mount service account token for grafana dashboard sidecar
After the kube-prometheus upgrade in #65, Grafana broke because the
k8s-sidecar container that loads custom dashboards needs access to
the service account token.  That access was removed in
prometheus-operator/kube-prometheus#1591 .

A quick search did not turn up a simple way to only mount the token in
one of the pod's containers (`k8s-sidecar`, not `grafana`), so we're
mounting it for all containers in the pod instead.  Might be possible to
tighten this up further.