music: install mopidy and snapcast-server for audio streaming
Allows streaming of music, podcasts, and internet radio stations
to [snapcast] clients, including one connected to a multi-channel
amplifier and built-in speakers.

Follow-up PRs will configure access to a music library on NAS
storage and an OPML list of podcast feeds.

The mopidy and snapcast-server images are built from [this repo]
by GitHub actions.

[snapcast]: https://mjaggard.github.io/snapcast/
[this repo]: https://github.com/bfritz/dockerfiles