gitlab-ce: decrease min and max replica counts for webservice pod
Update `HorizontalPodAutoscaler` settings for the GitLab webservice
pod to make it more likely for that pod to scale down to 1 replica.
Was annoying to always have 2 running for a small install as the
HTTP request logs are split between the two.

Reduced maximum count from 10 to 3 as well given the small size of
this cluster.