gitlab-ce: use sops to store GitLab rails secrets in git
GitLab encrypts several types of secrets, including runner tokens,
in its postgres database.  For k8s installations the keys to decrypt
those values live in in a k8s secret at `app-gitlab-ce-rails-secret`.

During cluster rebuilds, `app-gitlab-ce-rails-secret` is
reinitialized and the key to unlock tokens in the restored database
are lost.  Store `app-gitlab-ce-rails-secret` in git to prevent that.